+++ src/main/resources/static/index.html
@@ -12,7 +12,7 @@
 
   <div>
     <h1>Secure HTTP API - Spring Boot</h1>
-    <p class="lead">The mission expands on the REST API Level 0 Booster by securing a REST endpoint using <a
+    <p class="lead">The mission expands on the REST API Level 0 Example by securing a REST endpoint using <a
             href="https://access.redhat.com/products/red-hat-single-sign-on">Red Hat SSO</a>. Red Hat SSO implements the
       OAuth 2.0 specification and uses it to issue access tokens to provide clients with various access rights to secured
       resources. Securing an application with SSO enables you to add security to your applications while centralizing the
